server/requirements.txt
# ── Core Framework ────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.5.0

# ── Database ──────────────────────────────
motor==3.6.0
beanie==1.27.0

# ── Auth ──────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ── HTTP Client (for ML service) ─────────
httpx==0.27.0

# ── Utilities ─────────────────────────────
python-dotenv==1.0.1
python-multipart==0.0.12

# ── Dev & Testing ────────────────────────
pytest==8.3.0
pytest-asyncio==0.24.0
